refactor: Update file paths and components for our-team page
