+++ app/(home)/home.tsx
@@ -5,6 +5,7 @@ import Image from "next/image";
 import { motion } from "framer-motion";
 import Link from "next/link";
 import { ArrowTopRightIcon } from "@radix-ui/react-icons";
+import { SecondCall } from "./second-call";
 
 export function HomePage() {
   return (
@@ -45,16 +46,13 @@ export function HomePage() {
               High Impact
             </h1>
           </motion.div>
-
+          <SecondCall />
           <div className="flex flex-col md:flex-row justify-center items-center w-full text-secondary">
             <div className="text-2xl w-full flex flex-col gap-2 font-semibold border-l-8 border-secondary pl-3">
               <h2 className="font-bold md:text-3xl">Göttingen</h2>
               <h2>September 4th - 7th 2024</h2>
             </div>
-            <Link
-              href="/submission"
-              className="w-full"
-            >
+            <Link href="/submission" className="w-full">
               <div className="text-2xl flex flex-row gap-2 text-secondary items-center justify-center ring-2 ring-secondary group py-5 px-16 hover:text-primary2 hover:ring-primary2 transition duration-150 my-10 shadow-lg">
                 <h1 className="font-bold ">Submit Your Abstract</h1>
               </div>
