+++ app/(home)/second-call.tsx
@@ -1,0 +1,13 @@
+import { FileTextIcon } from "@radix-ui/react-icons";
+
+export function SecondCall() {
+  return (
+    <div className="flex flex-col items-center text-accent bg-white/70 py-10 gap-3">
+      <h1 className="text-5xl font-bold">SECOND CALL FOR ABSTRACT</h1>
+      <h2 className="text-4xl font-semibold">15th May - June 30th 2024</h2>
+      <div className="flex gap-2">
+        <p className="text-xl"> Submit your abstract now!</p>
+      </div>
+    </div>
+  );
+}
