--- app/about/committe-persons.tsx
@@ -1,299 +1,0 @@
-// board of directors
-import agnia from "@/assets/team/agnia.jpeg";
-import sekar from "@/assets/team/sekar.jpeg";
-import risma from "@/assets/team/risma.jpeg";
-import kemal from "@/assets/team/kemal.jpg";
-import gamma from "@/assets/team/gamma.jpeg";
-import geraldus from "@/assets/team/geraldus.jpeg";
-import caryn from "@/assets/team/caryn.jpg";
-
-// information technology
-import azhar from "@/assets/team/azhar.jpeg";
-import gerraldo from "@/assets/team/gerraldo.jpeg";
-import imel from "@/assets/team/imel.png";
-
-// Public Relations
-import aditya from "@/assets/team/aditya.jpg";
-import kafa from "@/assets/team/kafa.jpg";
-import bea from "@/assets/team/bea.jpg";
-import santi from "@/assets/team/santi.jpg";
-import kinasih from "@/assets/team/kinasih.jpg";
-import evelyn from "@/assets/team/evelyn.jpg";
-import syifa from "@/assets/team/syifa.jpg";
-import hapsari from "@/assets/team/hapsari.jpg";
-import fadlilah from "@/assets/team/fadlilahdila.jpg";
-
-// Program & Event
-import chaska from "@/assets/team/chaska.jpg";
-import hessel from "@/assets/team/hessel.jpg";
-import rohmaneo from "@/assets/team/rohmaneo.jpg";
-import eunike from "@/assets/team/eunike.png";
-import hosea from "@/assets/team/hosea.jpg";
-import marcelian from "@/assets/team/marcelian.jpg";
-import christine from "@/assets/team/christine.jpg";
-
-// Fundraising & Partnership
-import taqi from "@/assets/team/taqi.jpg";
-import nata from "@/assets/team/natadaffa.jpg";
-import maulana from "@/assets/team/maulana.jpg";
-import tasa from "@/assets/team/tasallia.jpg";
-import tazya from "@/assets/team/tazya.jpg";
-
-export const duaTeratas = [
-  {
-    imageSrc: sekar,
-    name: "Sekar Ayu Lestari",
-    position: "President",
-    positionPPIJ: "President of PPI Jerman",
-    education:
-      "M.Sc. Candidate Electrical and Electronics Engineering, Technische Universität München",
-    occupation: "Working Student in Measurement Engineering at IAB GmbH",
-  },
-  {
-    imageSrc: agnia,
-    name: "Agnia Dewi Larasati",
-    position: "Board of Advisory",
-    positionPPIJ: "President of PPI Jerman",
-    education:
-      "M.Sc. Candidate Electrical and Electronics Engineering, Technische Universität München",
-    occupation: "Working Student in Measurement Engineering at IAB GmbH",
-  },
-];
-
-export const tigaTeratas = [
-  {
-    imageSrc: gamma,
-    name: "Ignatius Gamma Tjandra",
-    position: "Secretary General",
-    education:
-      "B.Sc. Candidate Industrial Engineering and Management, Technische Universität Berlin",
-    occupation: "Working Student in Arbeitsvorbereitung, Stadler Deutschland",
-  },
-  {
-    imageSrc: geraldus,
-    name: "Geraldus Kevin Martimbang",
-    position: "Deputy of Director",
-    education:
-      "M.Sc. Candidate Regional Planning, Technische Universität München",
-    occupation: "M.A. Candidate Architecture, Technische Universität München",
-  },
-  {
-    imageSrc: caryn,
-    name: "Ivana Carynthia Ligo",
-    position: "Treasurer",
-    education:
-      "M.Sc. Candidate Sustainable International Agriculture, Georg-August-Universität Göttingen",
-    occupation: "Working Student in Data Engineering, HOMIRO Metrology GmbH",
-  },
-];
-
-export const scientificCommittee = [
-  {
-    imageSrc: risma,
-    name: "Risma Sari",
-    position: "Scientific Committee",
-    education:
-      "M.Sc. Candidate Sustainable International Agriculture, Georg-August-Universität Göttingen",
-    occupation: "Working Student in Data Engineering, HOMIRO Metrology GmbH",
-  },
-  {
-    imageSrc: kemal,
-    name: "Kemal Fadillah",
-    position: "Scientific Committee",
-    education:
-      "M.Sc. Candidate Sustainable International Agriculture, Georg-August-Universität Göttingen",
-    occupation: "Working Student in Data Engineering, HOMIRO Metrology GmbH",
-  },
-];
-
-export const teams = [
-  {
-    team: "Public Relations",
-    members: [
-      {
-        imageSrc: hapsari,
-        name: "Hapsari Pinji",
-        position: "Head of Public Relations, Design & Media",
-        education: "B.A.Sc. Angewandte Volkswirtschaftslehre, FU Berlin",
-        university: "FU Berlin",
-      },
-      {
-        imageSrc: fadlilah,
-        name: "Fadlilah Nur Hasanah",
-        position: "Staff of Public Relations, Design & Media",
-        education: "B.Sc. Bioinformatik",
-        univesrity: "Universität des Saarlandes",
-      },
-      {
-        imageSrc: bea,
-        name: "Nicole Beatrice",
-        position: "Staff of Public Relations, Design & Media",
-        education: "B.Sc. Candidate Architecture",
-        university: "TU Braunschweig",
-      },
-      {
-        imageSrc: santi,
-        name: "Santi Ananda Sari",
-        position: "Staff of Public Relations, Design & Media",
-        education: "Ausbildung Restaurantfachfrau",
-        university: "Berufsschule Hugo Junkers",
-      },
-      {
-        imageSrc: kinasih,
-        name: "Kinasih Sekarswara",
-        position: "Staff of Public Relations, Design & Media",
-        education:
-          "M.A. Candidate Development Economics and International Development",
-        university: "HTW Berlin",
-      },
-      {
-        imageSrc: aditya,
-        name: "Aditya Setyo Nugroho",
-        position: "Staff of Public Relations, Design & Media",
-        education: "Ausbildung Koch",
-        university: "The Grand Green Familux Resorts",
-      },
-      {
-        imageSrc: syifa,
-        name: "Syifa Perbowo",
-        position: "Staff of Public Relations, Design & Media",
-        education: "M.Sc. Management & Technology",
-        university: "TU München",
-      },
-      {
-        imageSrc: evelyn,
-        name: "Evelyn Sutiono",
-        position: "Staff of Public Relations, Design & Media",
-        education: "M.Sc. Molecular Medicine",
-        university: "Universität Göttingen",
-      },
-      {
-        imageSrc: kafa,
-        name: "Kafa Billahi Kafila",
-        position: "Staff of Public Relations, Design & Media",
-        education: "FSJ SHG Psychiatrie & Ergotherapie",
-        university: "",
-      },
-    ],
-  },
-  {
-    team: "Information Technology",
-    members: [
-      {
-        imageSrc: azhar,
-        name: "Azhar Rahadian",
-        position: "Head of Information Technology",
-        education: "B.Sc Informatics",
-        university: "TU Braunschweig",
-      },
-      {
-        imageSrc: gerraldo,
-        name: "Gerraldo F. A. Susilo",
-        position: "Staff of Information Technology",
-        education: "B.Sc Business Mathematics",
-        university: "HTW Berlin",
-      },
-      {
-        imageSrc: imel,
-        name: "Imelda Z. E. Putri",
-        position: "Staff of Information Technology",
-        education: "B.Sc Computer Science",
-        university: "Universität Osnabrück",
-      },
-    ],
-  },
-  {
-    team: "Program & Event",
-    members: [
-      {
-        imageSrc: chaska,
-        name: "Chaska R. Oscario",
-        position: "Head of Program & Event",
-        education: "B.Sc Management and Technology",
-        university: "TU München",
-      },
-      {
-        imageSrc: hessel,
-        name: "Hessel Juliust",
-        position: "Staff of Program & Event",
-        education: "PhD Computer Science",
-        university: "TU München",
-      },
-      {
-        imageSrc: christine,
-        name: "Christine Gracia Pratama",
-        position: "Staff of Program & Event",
-        education: "PhD Technology Innovation Management",
-        university: "TU Hamburg",
-      },
-      {
-        imageSrc: rohmaneo,
-        name: "M. Rohmaneo Daminto",
-        position: "Staff of Program & Event",
-        education: "PhD Geodesy and Geoinformatics",
-        university: "TU Berlin",
-      },
-      {
-        imageSrc: eunike,
-        name: "Eunika Amarya",
-        position: "Staff of Program & Event",
-        education: "M.A. English Language, Literature, and Culture",
-        university: "Universität Göttingen",
-      },
-      {
-        imageSrc: hosea,
-        name: "Hosea Simanjuntak",
-        position: "Staff of Program & Event",
-        education: "B.Sc Materials Engineering",
-        university: "TU Clausthal",
-      },
-      {
-        imageSrc: marcelian,
-        name: "Marcelian Grace",
-        position: "Staff of Program & Event",
-        education: "M.Sc Regional Planning",
-        university: "TU München",
-      },
-    ],
-  },
-  {
-    team: "Fundraising & Partnership",
-    members: [
-      {
-        imageSrc: taqi,
-        name: "M. Rafi Taqiyyuddiin",
-        position: "Head of Fundraising & Partnership",
-        education: "B.Sc Transport System",
-        university: "TU Berlin",
-      },
-      {
-        imageSrc: nata,
-        name: "Daffa Aushafi Nata",
-        position: "Staff of Fundraising & Partnership",
-        education: "B.A.Sc Industrial Engineering",
-        university: "Hochschule Darmstadt",
-      },
-      {
-        imageSrc: maulana,
-        name: "Muhammad Maulana",
-        position: "Staff of Fundraising & Partnership",
-        education: "M.A. National and International Administration Policy",
-        university: "Universität Postdam",
-      },
-      {
-        imageSrc: tazya,
-        name: "Iqlima K. Mumtazya Harmansyah",
-        position: "Staff of Fundraising & Partnership",
-        education: "B.Eng Applied Chemistry",
-        university: "Hochschule Kaiserslautern",
-      },
-      {
-        imageSrc: tasa,
-        name: "Tasallia Kamil",
-        position: "Staff of Fundraising & Partnership",
-        education: "B.Eng Mechatronics",
-        university: "Fachhochschule Aachen",
-      },
-    ],
-  },
-];
