+++ app/about/our-team/page.tsx
@@ -1,0 +1,5 @@
+import { Committee } from "./committe";
+
+export default function TeamPage() {
+  return <Committee />;
+}
