+++ app/about/page.tsx
@@ -10,41 +10,8 @@ import Link from "next/link";
 import { IconicOutput } from "./output-iconic";
 import { ArrowTopRightIcon } from "@radix-ui/react-icons";
 
-/* Commitee */
-import agnia from "@/assets/commitee/Agnia.jpeg";
-import sekar from "@/assets/commitee/sekar.jpeg";
-import gamma from "@/assets/commitee/gamma.jpeg";
-import geraldus from "@/assets/commitee/geraldus.jpeg";
-import caryn from "@/assets/commitee/caryn.jpg";
-import azhar from "@/assets/commitee/azhar.jpeg";
-import gerraldo from "@/assets/commitee/Gerraldo.jpeg";
-import imel from "@/assets/commitee/imel.png";
-import chaska from "@/assets/commitee/Chaska.jpg";
-import hessel from "@/assets/commitee/Hessel.jpg";
-import rohmaneo from "@/assets/commitee/Rohmaneo.jpg";
-import eunike from "@/assets/commitee/eunike.png";
-import hosea from "@/assets/commitee/Hosea.jpg";
-import marcelian from "@/assets/commitee/Marcelian.jpg";
-import hapsari from "@/assets/commitee/hapsari.jpg";
-import fadlilah from "@/assets/commitee/FadlilahDila.jpg";
-import bea from "@/assets/commitee/Bea.jpg";
-import santi from "@/assets/commitee/Santi.jpg";
-import kinasih from "@/assets/commitee/Kinasih.jpg";
-import evelyn from "@/assets/commitee/evelyn.jpg";
-import syifa from "@/assets/commitee/Syifa.jpg";
-import taqi from "@/assets/commitee/Taqi.jpg";
-import nata from "@/assets/commitee/NataDaffa.jpg";
-import maulana from "@/assets/commitee/Maulana.jpg";
-import tasa from "@/assets/commitee/Tasallia.jpg";
-import aditya from "@/assets/commitee/aditya.jpg";
-import kafa from "@/assets/commitee/Kafa.jpg";
-import tazya from "@/assets/commitee/tazya.jpg";
-import christine from "@/assets/commitee/Christine.jpg";
-import risma from "@/assets/commitee/Risma.jpeg";
-import kemal from "@/assets/commitee/Kemal.jpg";
-
 import nopic from "@/assets/img-round-placeholder.png";
-import { Committee } from "./committe";
+import { TeamCTA } from "./team-cta";
 
 export default function Page() {
   return (
@@ -116,34 +83,15 @@ export default function Page() {
             </p>
           </div>
         </div>
+        
         <IconicOutput />
+        <TeamCTA />
+
         <div className="w-full flex flex-col items-center justify-center">
           <h1 className="pt-10 pb-20 text-4xl font-semibold">Past ICONIC</h1>
           <IconicImages />
         </div>
       </div>
-
-      <Committee />
-      <a href="/reviewers">
-        <div className="mt-[0.5px] mb-[100px] text-2xl flex flex-row gap-2 text-secondary items-center justify-center ring-2 ring-secondary group w-fit py-5 px-10 hover:text-primary2 hover:ring-primary2 transition duration-150 my-10">
-          <h1 className="font-bold ">Reviewer Team of ICONIC 2024</h1>
-          <svg
-            width="15"
-            height="15"
-            viewBox="0 0 15 15"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-            className="w-6 h-6 opacity-0 group-hover:opacity-100 transition duration-75"
-          >
-            <path
-              d="M3.64645 11.3536C3.45118 11.1583 3.45118 10.8417 3.64645 10.6465L10.2929 4L6 4C5.72386 4 5.5 3.77614 5.5 3.5C5.5 3.22386 5.72386 3 6 3L11.5 3C11.6326 3 11.7598 3.05268 11.8536 3.14645C11.9473 3.24022 12 3.36739 12 3.5L12 9.00001C12 9.27615 11.7761 9.50001 11.5 9.50001C11.2239 9.50001 11 9.27615 11 9.00001V4.70711L4.35355 11.3536C4.15829 11.5488 3.84171 11.5488 3.64645 11.3536Z"
-              fill="currentColor"
-              fill-rule="evenodd"
-              clip-rule="evenodd"
-            ></path>
-          </svg>
-        </div>
-      </a>
     </main>
   );
 }
