--- app/our-team/page.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-export default function Page() {
-  return (
-    <main className="min-h-screen flex flex-col gap-4 items-center justify-center">
-      <h1 className="text-3xl md:text-5xl font-bold">Our Team</h1>
-      <h1 className="text-3xl md:text-5xl font-bold">TBA April 2024</h1>
-    </main>
-  );
-}
